+++ src/main/java/com/almostreliable/summoningrituals/compat/kubejs/binding/SummoningEntityBinding.java
@@ -1,24 +1,19 @@
 package com.almostreliable.summoningrituals.compat.kubejs.binding;
 
-import com.almostreliable.summoningrituals.compat.kubejs.builder.SummoningEntityBuilder;
+import com.almostreliable.summoningrituals.compat.kubejs.builder.SummoningEntityInputBuilder;
 import com.almostreliable.summoningrituals.compat.kubejs.builder.SummoningEntityOutputBuilder;
 
 import net.minecraft.core.Holder;
-import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.entity.EntityType;
 
 public interface SummoningEntityBinding {
 
-    static SummoningEntityBuilder input(Holder<EntityType<?>> entity) {
-        return new SummoningEntityBuilder(entity);
+    static SummoningEntityInputBuilder input(Holder<EntityType<?>> entity) {
+        return new SummoningEntityInputBuilder(entity);
     }
 
-    static SummoningEntityBuilder input(Holder<EntityType<?>> entity, int count) {
-        return new SummoningEntityBuilder(entity, count);
-    }
-
-    static SummoningEntityBuilder input(Holder<EntityType<?>> entity, int count, CompoundTag data) {
-        return new SummoningEntityBuilder(entity, count).data(data);
+    static SummoningEntityInputBuilder input(Holder<EntityType<?>> entity, int count) {
+        return new SummoningEntityInputBuilder(entity, count);
     }
 
     static SummoningEntityOutputBuilder output(Holder<EntityType<?>> entity) {