+++ src/main/java/com/almostreliable/summoningrituals/recipe/EntityInfo.java
@@ -2,7 +2,6 @@ package com.almostreliable.summoningrituals.recipe;
 
 import com.almostreliable.summoningrituals.core.Constants;
 
-import net.minecraft.core.BlockPos;
 import net.minecraft.core.Holder;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
@@ -20,12 +19,8 @@ import com.mojang.serialization.codecs.RecordCodecBuilder;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.function.Function;
-import java.util.function.Predicate;
 
-public record EntityInfo(
-    Holder<EntityType<?>> entity, int count, Optional<CompoundTag> data, List<Component> tooltip
-) implements Predicate<Entity> {
+public record EntityInfo(Holder<EntityType<?>> entity, int count, Optional<CompoundTag> data, List<Component> tooltip) {
 
     public static final Codec<EntityInfo> CODEC = RecordCodecBuilder.create(i -> i.group(
         BuiltInRegistries.ENTITY_TYPE.holderByNameCodec().fieldOf(Entity.ID_TAG).forGetter(EntityInfo::entity),
@@ -40,13 +35,4 @@ public record EntityInfo(
         ComponentSerialization.STREAM_CODEC.apply(ByteBufCodecs.list()), EntityInfo::tooltip,
         EntityInfo::new
     );
-
-    /**
-     * used for recipe matching, ignores data because data needs to be checked manually
-     * see {@link AltarRecipe#getSacrifices(BlockPos, Function)}
-     */
-    @Override
-    public boolean test(Entity entity) {
-        return entity.isAlive() && this.entity.value().equals(entity.getType());
-    }
 }